+++ graphics-engine/src/Engine/Level.h
@@ -6,6 +6,6 @@ class Level
 public:
 	Level() {};
 	virtual bool init();
-	virtual void update(float dt);
+	virtual void update(double dt);
 	virtual ~Level() {};
 };
