+++ graphics-engine/src/Engine/Time.h
@@ -3,11 +3,11 @@
 class Time
 {
 private:
-    static unsigned s_TimeCurrentFrame;
-    static unsigned s_TimeLastFrame;
-    static float deltaTime;
+    static double s_TimeCurrentFrame;
+    static double s_TimeLastFrame;
+    static double deltaTime;
 public:
     // Returns the time in miliseconds since the last frame
-    static float getDeltaTime() { return Time::deltaTime; }
+    static double getDeltaTime() { return Time::deltaTime; }
     static void tick();
 };
