+++ graphics-engine/src/Engine/glFramework/Levels/LevelRenderTirangle.h
@@ -12,7 +12,7 @@ class LevelRenderTriangle : public Level
 public:
 	LevelRenderTriangle();
 	bool init() override;
-	void update(float dt) override;
+	void update(double dt) override;
 	~LevelRenderTriangle();
 
 private:
